requirements.txt
altgraph==0.17
future==0.18.2
pefile==2019.4.18
Pillow==7.2.0
PyInstaller==3.6
PySimpleGUI==4.24.0
pywin32-ctypes==0.2.0
